[FEATURE] Add data-href as button attribute
Using <button> with href attribute in templates they are
checked by other software before the xsl mapping is executed
removing the href attribute from the button tag. Then an <a> tag
without an href attribute is generated.
We are using this inky Template in newsletter of a TYPO3 website.
There is an html-sanitizer which removes that href attribute.
Normally href is not an button attribute but data-* is no problem.
The patch adds data-href attribute to the xsl mapping.